# IGNITE-23083 Sql. Avoid planner cache pollution with simple insert queries