```
with optimization
Benchmark                         (clusterSize)  (fsync)  (partitionCount)  Mode  Cnt     Score    Error  Units
InsertBenchmark.sqlInlinedInsert              1    false                32  avgt   20  1011.028 ± 88.904  us/op no boost
Benchmark                         (clusterSize)  (fsync)  (partitionCount)  Mode  Cnt     Score    Error  Units
InsertBenchmark.sqlInlinedInsert              1    false                32  avgt   20  1024.413 ± 91.924  us/op no boost
Benchmark                         (clusterSize)  (fsync)  (partitionCount)  Mode  Cnt    Score    Error  Units
InsertBenchmark.sqlInlinedInsert              1    false                32  avgt   20  313.942 ± 56.033  us/op with boost

without
Benchmark                         (clusterSize)  (fsync)  (partitionCount)  Mode  Cnt     Score     Error  Units
InsertBenchmark.sqlInlinedInsert              1    false                32  avgt   20  1042.638 ± 151.594  us/op no boost
Benchmark                         (clusterSize)  (fsync)  (partitionCount)  Mode  Cnt     Score     Error  Units
InsertBenchmark.sqlInlinedInsert              1    false                32  avgt   20  1086.291 ± 196.949  us/op no boost
Benchmark                         (clusterSize)  (fsync)  (partitionCount)  Mode  Cnt    Score     Error  Units
InsertBenchmark.sqlInlinedInsert              1    false                32  avgt   20  362.045 ± 199.638  us/op with boost

with optimization
Benchmark                          (clusterSize)  (fsync)  (partitionCount)  Mode  Cnt    Score   Error  Units
InsertBenchmark.sqlPreparedInsert              1    false                32  avgt   20  117.460 ± 9.242  us/op

without 
Benchmark                          (clusterSize)  (fsync)  (partitionCount)  Mode  Cnt    Score    Error  Units
InsertBenchmark.sqlPreparedInsert              1    false                32  avgt   20  125.572 ± 35.950  us/op
```