@lowka 2024-08-29
I think this modification should not be more complex than

```
In prepareDml:

 return validFut.thenCompose(stmt -> {
            IgniteRel optInsert = tryOptimizeInsert(stmt.value.sqlNode(), ctx.planner());
            // Check key value Modify plan that its expression list only contains literals
            // and if that is the case return the `optInsert `.
            if (optInsert != null) {
                return optInsert;
            }

           ....

```

There is no need for cache invalidation.