@korlov42 2024-08-29
Looks like performance on insert became worse (I double-checked these results):

```
Before
Benchmark                          (clusterSize)  (fsync)  (partitionCount)  (replicaCount)  Mode  Cnt    Score   Error  Units
InsertBenchmark.sqlInlinedInsert               1    false                25               1  avgt   20  130.512 ± 2.422  us/op

After
Benchmark                         (clusterSize)  (fsync)  (partitionCount)  (replicaCount)  Mode  Cnt    Score   Error  Units
InsertBenchmark.sqlInlinedInsert              1    false                25               1  avgt   20  148.395 ± 2.009  us/op
```