@zstan 2024-08-30
```
opt
Benchmark                         (clusterSize)  (fsync)  (partitionCount)  (replicaCount)  Mode  Cnt    Score    Error  Units
InsertBenchmark.sqlInlinedInsert              1    false                 1               1  avgt   30  303.808 ± 40.896  us/op

no opt
Benchmark                         (clusterSize)  (fsync)  (partitionCount)  (replicaCount)  Mode  Cnt    Score    Error  Units
InsertBenchmark.sqlInlinedInsert              1    false                 1               1  avgt   30  286.148 ± 15.264  us/op
```